requirements.txt
Pillow==7.2.0
pytesseract==0.3.5
pdf2image==1.14.0